// src/features/suppliers/types/supplier.ts
export interface Supplier {
  id: string
  name: string
  contactName: string
  phone: string
  address: string
  isActive: boolean
  createdAt: string
  updatedAt: string
}

export interface SupplierResponse {
  id: string
  name: string
  contactName: string
  phone: string
  address: string
  isActive: boolean
  createdAt: string
  updatedAt: string
}